Completely change functionality because of environment variables
This change is completely incompatible with the previous version,
thus the bump in major version. If you are upgrading to this version
please note that config should not be uploaded using the included
cli and not as a document in S3.

This change utilizes (recently-added) native support for environment
variables in Lambda. It can thus eliminate its reliance (and thus
occasional slowness) on S3.